+++ extensions/aerospace/CHANGELOG.md
@@ -1,5 +1,14 @@
 # aerospace Changelog
 
+## [Improvements] - 2025-06-18
+
+- Adding logic for config path to match [AeroSpace Custom config location documentation](https://nikitabobko.github.io/AeroSpace/guide#config-location), using extension preferences first, then `~/.aerospace.toml`, then `${XDG_CONFIG_HOME}/aerospace/aerospace.toml`
+- Improving visual display of showShortcuts list
+  - Adding list sections by mode name
+  - Displaying multiple commands as "X & Y" instead of as an array "[X,Y]"
+  - Displaying keyboard shortcut keys as "tags" in list item's "accessories"
+    - Side note: this is a "workaround", as it does not display the keys like they are in the actions panel or in raycast root window
+
 ## [Bug Fixes] - 2025-01-10
 
 - Fix issue with `aerospace` not found if installed in a non-standard location (e.g. managed by `nix-darwin`)